+++ src/endpoints/basket/__tests__/createItem.test.ts
@@ -57,6 +57,28 @@ Object {
 `);
 });
 
+it('Support sending a child shop ID', () => {
+  expect(
+    createBasketItemRequest({
+      basketKey: 'basket1',
+      variantId: 1235,
+      quantity: 1,
+      childShopId: 456,
+    }),
+  ).toMatchInlineSnapshot(`
+Object {
+  "data": Object {
+    "quantity": 1,
+    "shopId": 456,
+    "variantId": 1235,
+  },
+  "endpoint": "baskets/basket1/items",
+  "method": "POST",
+  "params": Object {},
+}
+`);
+});
+
 it('Builds correct query with price promotion key', () => {
   expect(
     createBasketItemRequest({