+++ src/endpoints/wishlist/__tests__/addWishlistItem.test.ts
@@ -102,4 +102,24 @@ Object {
   },
 }
 `);
+
+  expect(
+    addWishlistItemEndpointRequest({
+      wishlistKey: 'wishlist_1',
+      item: {
+        productId: 1,
+      },
+      childShopId: 456,
+    }),
+  ).toMatchInlineSnapshot(`
+Object {
+  "data": Object {
+    "productId": 1,
+    "shopId": 456,
+  },
+  "endpoint": "wishlists/wishlist_1/items",
+  "method": "POST",
+  "params": Object {},
+}
+`);
 });