+++ src/endpoints/products/__tests__/productsByIds.test.ts
@@ -54,4 +54,26 @@ Object {
   },
 }
 `);
+
+  expect(() =>
+    createProductsByIdsEndpointRequest({
+      productIds: [],
+    }),
+  ).toThrow(`"productIds" parameter must not be an empty array.`);
+
+  expect(
+    createProductsByIdsEndpointRequest({
+      productIds: [1],
+      pricePromotionKey: 'abc123',
+    }),
+  ).toMatchInlineSnapshot(`
+Object {
+  "endpoint": "products",
+  "method": "GET",
+  "params": Object {
+    "ids": "1",
+    "pricePromotionKey": "abc123",
+  },
+}
+`);
 });