+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## UNRELEASED
+
+`createProductsByIdsEndpointRequest` throws when an empty array is provided for the `productIds` parameter.
+
+`BapiClient.products.getByIds` returns an empty array when an empty array is provided for the `productIds` parameter, not making any network request.
+
+`pricePromotionKey` can now be set for product listing request and products by IDs queries.
+
+Add `includeHidden` parameter to product's `categories` include to request hidden categories on the product.
+
+Support `childShopId` in `addWishlistItemEndpointRequest`.
+
 ## 1.5.0
 
 Set `accept-encoding: gzip, deflate` in NodeJS context